share configs
