+++ package.json
@@ -7,6 +7,8 @@
 		"cli:build": "tsc --build",
 		"cli-report:dev": "npm run dev --workspace=@cookie-analysis-tool/cli-report",
 		"cli-report:build": "npm run build --workspace=@cookie-analysis-tool/cli-report",
+		"extension:dev": "npm run dev --workspace=@cookie-analysis-tool/extension",
+		"extension:build": "npm run build --workspace=@cookie-analysis-tool/extension",
 		"cli":"node dist/cli/index.js",
 		"test": "jest --config=tests/jest.config.cjs",
 		"test:coverage": "npm run test -- --collectCoverage && open coverage/lcov-report/index.html",
