+++ packages/extension/babel.config.cjs
