+++ packages/extension/package.json
@@ -10,6 +10,10 @@
   "bugs": {
     "url": "https://github.com/GoogleChromeLabs/ps-analysis-tool/issues"
   },
+  "scripts": {
+    "dev": "webpack --config webpack.config.cjs --watch",
+		"build": "cross-env NODE_ENV=production webpack --config webpack.config.cjs"
+  },
   "homepage": "https://github.com/GoogleChromeLabs/ps-analysis-tool#readme",
   "dependencies": {
     "@tanstack/react-table": "^8.9.3",
