+++ postcss.config.cjs
@@ -19,7 +19,7 @@ const path = require('path');
 module.exports = {
   plugins: {
     tailwindcss: {
-      config: path.join(__dirname, './packages/extension/tailwind.config.cjs'),
+      config: path.join(__dirname, 'tailwind.config.cjs'),
     },
     autoprefixer: {},
   },
